+++ archive/cli/create.py
@@ -5,21 +5,7 @@ from pathlib import Path
 from archive.archive import Archive, DedupMode
 
 
-suffix_map = {
-    '.tar': 'none',
-    '.tar.gz': 'gz',
-    '.tar.bz2': 'bz2',
-    '.tar.xz': 'xz',
-}
-"""Map path suffix to compression mode."""
-
 def create(args):
-    if args.compression is None:
-        try:
-            args.compression = suffix_map["".join(args.archive.suffixes)]
-        except KeyError:
-            # Last ressort default
-            args.compression = 'gz'
     if args.compression == 'none':
         args.compression = ''
     archive = Archive().create(args.archive, args.compression, args.files,